+++ src/commands/experimental/yaml/yaml.test.ts
@@ -3,7 +3,7 @@ import GenerateIntegrationFromYAMLCommand from "./index.js";
 import fs from "fs";
 import path from "path";
 import { readFile } from "fs-extra";
-import { walkDir } from "../../../../fs.js";
+import { walkDir } from "../../../fs.js";
 
 const FLOW_GENERATION_TIMEOUT = 5 * 60 * 1000; // 5 minutes
 interface SpecMeta {
@@ -13,19 +13,17 @@ interface SpecMeta {
 
 describe("YAML CNI generation tests", () => {
   const basePath = process.cwd();
-  const integrationsPath = path.resolve("src/commands/integrations");
-  const tempPath = path.resolve(`${integrationsPath}/init/yaml/temp`);
+  const commandPath = path.resolve("src/commands/experimental/yaml");
+  const tempPath = path.resolve(`${commandPath}/temp`);
 
   if (!fs.existsSync(tempPath)) {
     fs.mkdirSync(tempPath);
   }
 
-  const specs = fs
-    .readdirSync(`${integrationsPath}/init/yaml/fixtures/specs`)
-    .map<SpecMeta>((fileName) => {
-      const [name] = fileName.toLowerCase().split(".");
-      return { fileName, name };
-    });
+  const specs = fs.readdirSync(`${commandPath}/fixtures/specs`).map<SpecMeta>((fileName) => {
+    const [name] = fileName.toLowerCase().split(".");
+    return { fileName, name };
+  });
 
   for (const { fileName, name } of specs) {
     describe(name, () => {
